move filter functions to own class
